chore: env agnostic contract loader supp both node and browser with File API
